requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
transformers>=4.36.0
torch>=2.1.0
pillow>=10.0.0
matplotlib>=3.7.0
plotly>=5.18.0
numpy>=1.24.0
pyyaml>=6.0.0
slack-bolt
slack-sdk
